Ask for the user's location after he/she uploaded the photo.
This way I think it's less creepy than seeing the message as soon
as you enter the site for the first time. To do so, I had to move
the geolocation.js to the show page. But I only load it if the user
just uploaded the picture, so the geolocation doesn't get set if
someone else opens that URL.

It might be cleaner to check if the photo already has geolocation and,
if not, ask for the user's location.

I leave this as an exercise for the reader :)